backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
distro==1.9.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
lxml==4.9.1
numpy==1.23.4
openai==1.34.0
openpyxl==3.0.10
pandas==1.5.1
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pypdfium2==4.30.0
python-dateutil==2.9.0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.64.1
typing_extensions==4.12.2
urllib3==2.2.3
XlsxWriter==3.0.9
